Move current graphic picker into "basic"
